requirements.txt
alembic==1.12.1
annotated-types==0.6.0
anyio==3.7.1
bcrypt==3.2.0
cffi==1.17.1
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.104.1
greenlet==3.0.1
h11==0.14.0
httptools==0.6.1
idna==3.4
Mako==1.3.0
MarkupSafe==2.1.3
passlib==1.7.4
psycopg2-binary==2.9.9
pyasn1==0.5.1
pycparser==2.22
pydantic==2.5.1
pydantic-settings==2.1.0
pydantic_core==2.14.3
PyMySQL==1.1.1
pyotp==2.9.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.24.0.post1
watchfiles==0.21.0
websockets==12.0
